Regression + Models

+++ src/SensorDataset.py
@@ -74,6 +74,7 @@ class SensorDataset():
             self.x_test, self.y_test = self.split_windows(group_size=group_size,
                                                           step_size=step_size, X=self.x_test, y=self.y_test)
 
+
         print("Train {}.{}".format(self.x_train.shape, self.y_train.shape))
         print("Test {}.{}".format(self.x_test.shape, self.y_test.shape))
 
@@ -101,7 +102,7 @@ class SensorDataset():
         return np.eye(n_values)[np.array(y_, dtype=np.int32)]  # Returns FLOATS
 
     def split_windows(self, group_size, step_size, X, y):
-        number_groups =((X.shape[1]-group_size)/step_size)+1
+        number_groups = int(((X.shape[1]-group_size)/step_size)+1)
         split_xy = [(X[j, i:i + group_size], y[j]) for j in range(len(X)) for i in range(0, number_groups * step_size, step_size)]
         split_x = np.array([x[0] for x in split_xy])
         split_y = np.array([y[1] for y in split_xy])
